refactor(auth): keep browser filter private
